article type별도 파일 분리
