+++ src/utils/api.ts
@@ -1,4 +1,5 @@
 import axios from "axios";
+import { Welcome } from "types/article";
 
 export const getNews = async () => {
   const response = await axios.get<Welcome>(
@@ -6,22 +7,3 @@ export const getNews = async () => {
   );
   return response.data.articles;
 };
-export interface Welcome {
-  articles: Article[];
-}
-
-export interface Article {
-  source: Source;
-  author: null | string;
-  title: string;
-  description: string;
-  url: string;
-  urlToImage: string;
-  publishedAt: Date;
-  content: null | string;
-}
-
-export interface Source {
-  id: null;
-  name: string;
-}
