+++ types/article.ts
@@ -1,0 +1,19 @@
+export interface Welcome {
+  articles: Article[];
+}
+
+export interface Article {
+  source: Source;
+  author: null | string;
+  title: string;
+  description: string;
+  url: string;
+  urlToImage: string;
+  publishedAt: Date;
+  content: null | string;
+}
+
+export interface Source {
+  id: null;
+  name: string;
+}
